requirements.txt
sounddevice==0.3.14
tenacity==4.12.0
Django==2.2.6
grpcio==1.24.3
pathlib2==2.3.5
click==6.7
protobuf==3.10.0
grpc==0.3-19
djangorestframework==3.10.3
